Generate SSH config for all nodes
We only generated one for the bootstrap node, but we now also need to
SSH into the router node.

A simple Bash script is introduced to generate the SSH config file (may
be rewritten in Python when we extract this Terraform tooling for use
outside of our CI context).

Issue: GH-889