Ignore node IPs in Terraform output
We wanted to group all IP addresses under a single output variable, so
that one could `terraform output ips` to get a clear view of what was
spawned (and so we could generate a SSH config file for accessing any of
the spawned VMs by name).

However, the "splat" syntax we wanted to use is not supported in
Terraform <= 0.11, so we just ignore the nodes (other than bootstrap and
router) for now.

For reference: https://github.com/hashicorp/terraform/issues/17048

Issue: GH-889