Wait for Pods to be Running before reading logs
We had numerous occurences of this test failing due to some Pod in the
list still `Creating`. We now wait before attempting to read logs.

This band-aid is not supposed to remain here for long. The goal is to
remove this test, and replace it with a more useful one: check that when
running 'echo something' in a Pod, we can retrieve it afterwards through
the Pod logs.