kubernetes: Add method to obtain oidc service IP
We need to be able to map the OIDC service to a static clusterIP address.
The OIDC service IP is choosen as the 7th ip-address in the network.

Initially, the OIDC service was exposed using an ingress object hence no
need for a dedicated service IP but this approach was later rejected
since the ingress does not perform valid tls termination.