ci: Introduce library to generate CI build plan
In order to ease the management of "eve/main.yml", we want to generate
it from code.

This commit introduces a simple Python library to generate some of the
YAML objects often used in this "eve/main.yml" file. The generation
relies on `OrderedDict` to ensure consistent results across executions.

We also setup unit tests from the start, and provide `tox` entrypoints
for both running those tests and running the buildplan generation (to be
implemented in the future).