ui/salt: Allow arbitrary props in Job objects
We previously used a single prop, `name`, to encode information about
the Job at hand. For Node deployment Jobs, we used the
`deploy-node/<nodeName>` pattern, which is cumbersome to manipulate.

Since this is JSON, there is no overhead to add optional dimensions, and
there will be close to no impact on localStorage size.