salt, kubernetes: remove `apiServer` from `BootstrapConfiguration`
We no longer need this since we provide in-cluster HA for
`kube-apiserver` access. If this is desired for out-of-cluster access,
we can provide this using a `LoadBalancer` `Service` once we have the
infrastructure to support this in place.

This also removed the optional deployment of `keepalived`.

See: https://github.com/scality/metalk8s/issues/2103
See: https://github.com/scality/metalk8s/issues/1788